refactor: 去掉多余的 _Identity, 路径解析直接填 FileInfo
