+++ src/amane/parsing/__init__.py
@@ -1,11 +1,10 @@
 from .actor_name import split_actor_aliases
-from .file_info import FileInfo, parse_file_info
+from .file_info import FileInfo, infer_content_type, parse_file_info
 from .number import (
     ContentType,
     classify_number,
     extract_number,
     get_prefix,
-    infer_content_type,
     is_amateur,
     is_uncensored,
 )
